# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library API",
    "description": "REST API for managing a library — books and authors collections. Protected routes require GitHub OAuth login at /auth/github.",
    "version": "1.0.0"
  },
  "host": "cse341-librarymanager.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "securityDefinitions": {
    "githubOAuth": {
      "type": "oauth2",
      "authorizationUrl": "https://cse341-librarymanager.onrender.com/auth/github",
      "flow": "implicit",
      "scopes": {
        "user:email": "Read user email from GitHub"
      }
    }
  },
  "paths": {
...
    "/auth/github/callback": {
      "get": {
        "tags": ["Authentication"],
        "summary": "GitHub OAuth callback",
        "description": "GitHub redirects here after authentication. On success, redirects to /auth/profile.",
        "responses": {
          "302": { "description": "Redirect to /auth/profile on success" },
          "401": { "description": "Authentication failed" }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get logged-in user profile",
        "description": "Returns the currently authenticated user's GitHub profile.",
        "responses": {
          "200": { "description": "User profile returned" },
          "401": { "description": "Not authenticated" }
        }
      }
    },
...
  }
}